% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_connect}
\alias{dm_connect}
\title{Connects to the data manager and returns an auth token}
\usage{
dm_connect(email, password)
}
\arguments{
\item{email}{character, email address for data manager}

\item{password}{character, password for data manager}
}
\value{
A list of characters containing the url_base and auth token
}
\description{
Connects to the data manager and returns an auth token
}
