% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_download_file}
\alias{dm_download_file}
\title{Using the download link from dm_get_dataset_status(), downloads data
files to the specified local path}
\usage{
dm_download_file(presigned_url, output_path)
}
\arguments{
\item{presigned_url, }{character. 'zip' or 'pq' ONLY.}

\item{output_path, }{character}
}
\description{
Using the download link from dm_get_dataset_status(), downloads data
files to the specified local path
}
