% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_generate_dataset}
\alias{dm_generate_dataset}
\title{Requests a dataset to be generated using the requested parameters}
\usage{
dm_generate_dataset(
  con,
  project_name,
  scenarios,
  family_variables = list(),
  person_variables = list(),
  birth_year_range = list(),
  year_range = list()
)
}
\arguments{
\item{con}{A list of connection information}

\item{project_name, }{character}

\item{scenarios, }{list of characters}

\item{family_variables, }{list of characters}

\item{person_variables, }{list of characters}

\item{birth_year_range, }{list of integers}

\item{year_range, }{list of integers}
}
\value{
a list of characters containing a job_id and message
}
\description{
Requests a dataset to be generated using the requested parameters
}
