% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_get_dataset_status}
\alias{dm_get_dataset_status}
\title{Retrieves the status of the job_id returned from dm_generate_dataset().
If the dataset has successfully generated, dm_get_dataset_status() will
return download links to family and person data files.}
\usage{
dm_get_dataset_status(job_id, file_type)
}
\arguments{
\item{job_id, }{integer}

\item{file_type, }{character. 'csv' or 'parquet' ONLY}
}
\value{
a list
}
\description{
Retrieves the status of the job_id returned from dm_generate_dataset().
If the dataset has successfully generated, dm_get_dataset_status() will
return download links to family and person data files.
}
