% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_get_project}
\alias{dm_get_project}
\title{Retrieves a single project using its project_id}
\usage{
dm_get_project(con, project_id)
}
\arguments{
\item{con}{A list of connection information}
}
\value{
A dataframe of project information
}
\description{
Retrieves a single project using its project_id
}
