% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_get_projects}
\alias{dm_get_projects}
\title{Retrieves all projects in the data manager}
\usage{
dm_get_projects()
}
\value{
A dataframe of projects
}
\description{
Retrieves all projects in the data manager
}
