% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_get_scenarios}
\alias{dm_get_scenarios}
\title{Retrieves all scenarios}
\usage{
dm_get_scenarios(con)
}
\arguments{
\item{con}{A list of connection information}
}
\value{
A dataframe of scenarios
}
\description{
Retrieves all scenarios
}
