% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_get_variables_for_project}
\alias{dm_get_variables_for_project}
\title{Retrieves the variables for a project}
\usage{
dm_get_variables_for_project(con, project_id)
}
\arguments{
\item{con}{A list of connection information}

\item{project_id, }{integer identifier for a project}
}
\value{
a dataframe containing family and person variables as dataframes
}
\description{
Retrieves the variables for a project
}
