% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_request_and_download_datasets}
\alias{dm_request_and_download_datasets}
\title{Requests and downloads a dataset using dm_generate_dataset(),
dm_get_dataset_status(), and dm_download_file}
\usage{
dm_request_and_download_datasets(
  con,
  file_type,
  output_dir,
  project_name,
  scenarios,
  family_variables = list(),
  person_variables = list(),
  birth_year_range = list(),
  year_range = list()
)
}
\arguments{
\item{con}{A list of connection information}

\item{file_type, }{'csv' or 'parquet' ONLY}

\item{output_dir, }{character}

\item{project_name, }{character}

\item{scenarios, }{list of characters}

\item{family_variables, }{list of characters}

\item{person_variables, }{list of characters}

\item{birth_year_range, }{list of integers}

\item{year_range, }{list of integers}
}
\description{
Requests and downloads a dataset using dm_generate_dataset(),
dm_get_dataset_status(), and dm_download_file
}
