% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{feh_read}
\alias{feh_read}
\title{Reads FEH header and data files}
\usage{
feh_read(
  fehdir,
  suffix = "even",
  person = TRUE,
  columns = NULL,
  obs_count = NA
)
}
\arguments{
\item{fehdir}{the folder containing the header and data files}

\item{suffix}{(optional) "even", "odd", or NULL}

\item{person}{(optional)}

\item{columns}{(optional) character vector of column names to keep; all if NULL (default)}

\item{obs_count}{the number of observations to read; all if NA (default)}
}
\value{
data frame with FEH data
}
\description{
File names must follow the following convention:
header: dynasipp_header_[even|odd].dat
data:   dynasipp_[person|family]_[even|odd].dat
}
