% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_feh_input_files.R
\name{feh_read_alters}
\alias{feh_read_alters}
\title{Reads ALTER statements from rundynasipp.inp into a data frame.}
\usage{
feh_read_alters(fname)
}
\arguments{
\item{fname}{path for a rundynasipp.inp file}
}
\value{
dataframe with columns "TSNAME","YEAR","VALUE"
}
\description{
Reads ALTER statements from rundynasipp.inp into a data frame.
}
