% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_feh_input_files.R
\name{feh_update_alters}
\alias{feh_update_alters}
\title{Creates a new version of input file with updated ALTER statements}
\usage{
feh_update_alters(alterdf, infname, outfname)
}
\arguments{
\item{alterdf}{dataframe with new ALTERs and columns: "TSNAME", "YEAR", "VALUE"}

\item{infname}{initial input file path}

\item{outfname}{the path of the new input file}
}
\value{
NA
}
\description{
Creates a new version of input file with updated ALTER statements
}
