% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{feh_wide_to_long}
\alias{feh_wide_to_long}
\title{Converts DYNASIM FEH output data from wide to long format.
It also changes the column names into lower case.}
\usage{
feh_wide_to_long(data)
}
\arguments{
\item{data}{DYNASIM FEH dataset as returned by read_feh}
}
\value{
a dataframe in long format
}
\description{
Converts DYNASIM FEH output data from wide to long format.
It also changes the column names into lower case.
}
