% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_feh_input_files.R
\name{feh_write_new_input}
\alias{feh_write_new_input}
\title{Writes a new input file with updated ALTER statements}
\usage{
feh_write_new_input(lines, df, fname)
}
\arguments{
\item{lines}{character vector with lines from the original input files}

\item{df}{dataframe with columns: "TSNAME", "YEAR", "VALUE"}

\item{fname}{the name of the new input file}
}
\value{
NA
}
\description{
Writes a new input file with updated ALTER statements
}
