% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{get_record_types}
\alias{get_record_types}
\title{Returns a vector of field types}
\usage{
get_record_types(lines)
}
\arguments{
\item{lines}{Character vector of codebook lines}
}
\value{
A vector of field types
}
\description{
Returns a vector of field types
}
