% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{make_col_names}
\alias{make_col_names}
\title{Creates a vector of column names from a header record}
\usage{
make_col_names(rec)
}
\arguments{
\item{rec}{record structure returned by make_record_struct}
}
\description{
Creates a vector of column names from a header record
}
