% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{parse_codebook_lines}
\alias{parse_codebook_lines}
\title{Parse codebook lines into a dataframe of names, types and array ranges}
\usage{
parse_codebook_lines(lines)
}
\arguments{
\item{lines}{Character vector of codebook lines}
}
\value{
A dataframe with variables: name, type, start, stop
}
\description{
Parse codebook lines into a dataframe of names, types and array ranges
}
