% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{parse_family_codebook}
\alias{parse_family_codebook}
\title{Parse family record codebook}
\usage{
parse_family_codebook(fname)
}
\arguments{
\item{fname}{The codebook file name}
}
\value{
A dataframe with variables: name, type, start, stop
}
\description{
Parse family record codebook
}
