% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{parse_person_codebook}
\alias{parse_person_codebook}
\title{Parse person record codebook}
\usage{
parse_person_codebook(fname)
}
\arguments{
\item{fname}{The codebook file name}
}
\value{
A dataframe with variables: name, type, start, stop
}
\description{
Parse person record codebook
}
