% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_feh_input_files.R
\name{print_diff}
\alias{print_diff}
\title{Compares the updated and initial dataframes with ALTERs}
\usage{
print_diff(df)
}
\arguments{
\item{df}{dataframe with columns: "TSNAME", "YEAR", "VALUE.x", "VALUE.y", "VALUE"}
}
\value{
NA
}
\description{
Compares the updated and initial dataframes with ALTERs
}
