% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feh_micro.R
\name{read_feh_cbk}
\alias{read_feh_cbk}
\title{Reads DYNASIM codebook and FEH persons file}
\usage{
read_feh_cbk(codebookfile, fehfile, columns = NULL, obs_count = NA)
}
\arguments{
\item{codebookfile}{the path to an FEH codebook}

\item{fehfile}{the path to an FEH persons file}

\item{columns}{(optional) character vector of column names to keep; all if NULL (default)}

\item{obs_count}{the number of observations to read; all if NA (default)}
}
\description{
Reads DYNASIM codebook and FEH persons file
}
