% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager_main.R
\name{dm_get_scenarios_for_project}
\alias{dm_get_scenarios_for_project}
\title{Retrieves the scenarios for a project}
\usage{
dm_get_scenarios_for_project(project_id)
}
\arguments{
\item{project_id, }{integer identifier for a project}
}
\value{
A dataframe of scenarios
}
\description{
Retrieves the scenarios for a project
}
